# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Types.ObjectId": {
				"type": "string"
			},
			"IUser": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"gender": {
						"type": "string",
						"enum": [
							"male",
							"female"
						]
					},
					"DOB": {
						"type": "string",
						"format": "date-time"
					},
					"address": {
						"items": {
							"properties": {
								"manipolicity": {
									"type": "string"
								},
								"government": {
									"type": "string"
								},
								"country": {
									"type": "string"
								}
							},
							"required": [
								"manipolicity",
								"government",
								"country"
							],
							"type": "object"
						},
						"type": "array"
					},
					"verified": {
						"type": "string",
						"enum": [
							"notSent",
							"pending",
							"verified"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"inActive",
							"active",
							"suspended",
							"lost",
							"deleted"
						]
					},
					"accountType": {
						"type": "string",
						"enum": [
							"PT",
							"EM",
							"PA"
						]
					},
					"lastLoginDate": {
						"type": "string",
						"format": "date-time"
					},
					"accountSetting": {
						"additionalProperties": false,
						"type": "object"
					},
					"languages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"maritalStatus": {
						"type": "string",
						"enum": [
							"married",
							"single",
							"divorced",
							"widow"
						]
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"firstName",
					"lastName",
					"gender",
					"DOB",
					"verified",
					"status",
					"accountType",
					"maritalStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IServiceProvider": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"uID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUser"
							}
						]
					},
					"bio": {
						"type": "string"
					},
					"specialties": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"preferredServiceType": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"type": "string",
									"enum": [
										"online",
										"home",
										"office"
									]
								},
								"type": "array"
							}
						]
					},
					"minSessionFee": {
						"type": "number",
						"format": "double"
					},
					"maxSessionFee": {
						"type": "number",
						"format": "double"
					},
					"documents": {
						"items": {
							"properties": {
								"type": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"type",
								"name",
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"reviewerUIDs": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IUser"
								},
								"type": "array"
							}
						]
					},
					"verificationStatus": {
						"type": "string",
						"enum": [
							"notSubmitted",
							"pendingReview",
							"inReview",
							"verified",
							"rejected"
						]
					},
					"verificationDate": {
						"type": "string",
						"format": "date-time"
					},
					"verifiedByUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUser"
							}
						]
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"specialties",
					"preferredServiceType",
					"verificationStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IServiceProvider_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"uID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUser"
							}
						]
					},
					"bio": {
						"type": "string"
					},
					"specialties": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"preferredServiceType": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"type": "string",
									"enum": [
										"online",
										"home",
										"office"
									]
								},
								"type": "array"
							}
						]
					},
					"minSessionFee": {
						"type": "number",
						"format": "double"
					},
					"maxSessionFee": {
						"type": "number",
						"format": "double"
					},
					"documents": {
						"items": {
							"properties": {
								"type": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"type",
								"name",
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"reviewerUIDs": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Types.ObjectId"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IUser"
								},
								"type": "array"
							}
						]
					},
					"verificationStatus": {
						"type": "string",
						"enum": [
							"notSubmitted",
							"pendingReview",
							"inReview",
							"verified",
							"rejected"
						]
					},
					"verificationDate": {
						"type": "string",
						"format": "date-time"
					},
					"verifiedByUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUser"
							}
						]
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"Partial_IUser_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"gender": {
						"type": "string",
						"enum": [
							"male",
							"female"
						]
					},
					"DOB": {
						"type": "string",
						"format": "date-time"
					},
					"address": {
						"items": {
							"properties": {
								"manipolicity": {
									"type": "string"
								},
								"government": {
									"type": "string"
								},
								"country": {
									"type": "string"
								}
							},
							"required": [
								"manipolicity",
								"government",
								"country"
							],
							"type": "object"
						},
						"type": "array"
					},
					"verified": {
						"type": "string",
						"enum": [
							"notSent",
							"pending",
							"verified"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"inActive",
							"active",
							"suspended",
							"lost",
							"deleted"
						]
					},
					"accountType": {
						"type": "string",
						"enum": [
							"PT",
							"EM",
							"PA"
						]
					},
					"lastLoginDate": {
						"type": "string",
						"format": "date-time"
					},
					"accountSetting": {
						"additionalProperties": false,
						"type": "object"
					},
					"languages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"maritalStatus": {
						"type": "string",
						"enum": [
							"married",
							"single",
							"divorced",
							"widow"
						]
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IRole": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"employees": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"users": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"service_provider": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"clients": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"sessions": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"communications": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"disputes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"enum_values": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"employees",
					"users",
					"clients",
					"sessions",
					"communications",
					"disputes",
					"enum_values"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IRole_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"employees": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"users": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"service_provider": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"clients": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"sessions": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"communications": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"disputes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"enum_values": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"Document_any.any.any_": {
				"$ref": "#/components/schemas/FlattenMaps_T_",
				"description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
			},
			"Required_any_": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T required"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/serviceProviders/{serviceProviderId}": {
			"get": {
				"operationId": "GetServiceProvider",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IServiceProvider"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested serviceProvider is not found"
					}
				},
				"description": "Get a serviceProvider details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "serviceProviderId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63107aae6bcd70f0f363847c"
					}
				]
			},
			"delete": {
				"operationId": "DeleteServiceProvider",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IServiceProvider"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested serviceProvider is not found"
					}
				},
				"description": "Delete a serverProvider",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "serviceProviderId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63107aae6bcd70f0f363847c"
					}
				]
			}
		},
		"/serviceProviders/create": {
			"post": {
				"operationId": "CreateServiceProvider",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IServiceProvider"
								},
								"examples": {
									"Example 1": {
										"value": {
											"uID": "63124dffdcf1e4974079a441",
											"bio": "this is the best",
											"specialties": [
												"Physiotherapist"
											],
											"preferredServiceType": [
												"online",
												"home",
												"office"
											],
											"minSessionFee": 100,
											"maxSessionFee": 150,
											"documents": [
												{
													"url": "http://localhost:8000/documents",
													"name": "certificate",
													"type": "pdf"
												}
											],
											"reviewerUIDs": [
												"63124dffdcf1e4974079a441",
												"63130ba28d796320172e58d6"
											],
											"verificationStatus": "notSubmitted",
											"verificationDate": "2022-09-10T00:00:00.000Z",
											"verifiedByUID": "63130ba28d796320172e58d6"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a serviceProvider",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IServiceProvider"
							}
						}
					}
				}
			}
		},
		"/serviceProviders/update/{serviceProviderId}": {
			"put": {
				"operationId": "UpdateServiceProvider",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IServiceProvider"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a ServiceProvider",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "serviceProviderId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63107aae6bcd70f0f363847c"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IServiceProvider_"
							},
							"example": "63107aae6bcd70f0f363847c"
						}
					}
				}
			}
		},
...
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IUser"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested user is not found"
					}
				},
				"description": "Get a user details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63124dffdcf1e4974079a441"
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IUser"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested user is not found"
					}
				},
				"description": "Delete a user",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63124dffdcf1e4974079a441"
					}
				]
			}
		},
		"/users/create": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								},
								"examples": {
									"Example 1": {
										"value": {
											"email": "nour@gmail.com",
											"phoneNumber": "00352681531905",
											"password": "00352681531905",
											"profilePicture": "main.png",
											"firstName": "Muhammad",
											"lastName": "Nour",
											"gender": "male",
											"DOB": "2022-09-10T00:00:00.000Z",
											"address": [
												{
													"country": "Syria",
													"government": "SY",
													"manipolicity": "Fixed"
												}
											],
											"verified": "notSent",
											"status": "active",
											"accountType": "EM",
											"lastLoginDate": "2022-09-10T00:00:00.000Z",
											"accountSetting": {
												"name": "Setting1",
												"value": "Some value"
											},
											"languages": [
												"AR",
												"EN"
											],
											"maritalStatus": "single"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation failed"
					}
				},
				"description": "Create a user",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUser"
							}
						}
					}
				}
			}
		},
		"/users/update/{userId}": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IUser"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update user",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63124dffdcf1e4974079a441"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IUser_"
							},
							"example": "63124dffdcf1e4974079a441"
						}
					}
				}
			}
		},
...
		"/roles/{roleId}": {
			"get": {
				"operationId": "GetRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IRole"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested role is not found"
					}
				},
				"description": "Get role details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "631312818d796320172e58e5"
					}
				]
			},
			"delete": {
				"operationId": "DeleteRole",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IRole"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested role is not found"
					}
				},
				"description": "Delete role",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "631312818d796320172e58e5"
					}
				]
			}
		},
		"/roles/create": {
			"post": {
				"operationId": "CreateRole",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRole"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed"
					}
				},
				"description": "Create role",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRole"
							}
						}
					}
				}
			}
		},
		"/roles/update/{roleId}": {
			"put": {
				"operationId": "UpdateRole",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IRole"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update role",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "631312818d796320172e58e5"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IRole_"
							},
							"example": "631312818d796320172e58e5"
						}
					}
				}
			}
		},
...
		"/sessions/{sessionId}": {
			"get": {
				"operationId": "GetSession",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ISession"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested session in not found"
					}
				},
				"description": "Get a Session details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sessionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteSession",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/ISession"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested session in not found"
					}
				},
				"description": "Delete a session",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sessionId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63131c361ea32f520cb28bd6"
					}
				]
			}
		},
		"/sessions/create": {
			"post": {
				"operationId": "CreateSession",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISession"
								},
								"examples": {
									"Example 1": {
										"value": {
											"sessionType": "individual",
											"serviceProviderID": "63107aae6bcd70f0f363847c",
											"clientsID": [
												"6310750be8f4ab035351fb78",
												"63107517e8f4ab035351fb7a"
											],
											"name": "math",
											"details": "mathematical analysis",
											"startDate": "2022-09-10T00:00:00.000Z",
											"duration": 50,
											"serviceType": "home",
											"location": {
												"city": "Aleppo"
											},
											"attachments": [
												{
													"url": "www.google.com",
													"name": "cd",
													"type": "pdf"
												}
											],
											"requirements": "no requirement",
											"ratings": [
												{
													"raterUID": "6310750be8f4ab035351fb78",
													"ratingValue": "5",
													"ratingDate": "2022-09-10T00:00:00.000Z"
												}
											],
											"reviews": [
												{
													"reviewerUID": "6310750be8f4ab035351fb78",
													"reviewDetails": "goooood",
													"reviewDate": "2022-09-10T00:00:00.000Z"
												}
											],
											"sessionFee": 1500,
											"payments": [
												{
													"discount": 25,
													"paymentMethod": "paypal",
													"payerID": "6310750be8f4ab035351fb78",
													"amount": 250
												}
											],
											"status": "agreed",
											"doctorReferral": "no "
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a session",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISession"
							}
						}
					}
				}
			}
		},
		"/sessions/update/{sessionId}": {
			"put": {
				"operationId": "UpdateSession",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ISession"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a session",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sessionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ISession_"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}